added ore log mesages.

+++ kernel/include/interupts.h
@@ -23,7 +23,7 @@
 #define CASCADE_IRQ         2
 
 void io_wait(void);
-void PIC_remap(int offset1, int offset2);
+void PIC_remap(uint8_t offset1, uint8_t offset2);
 void PIC_send_EOI(uint8_t irq);
 void IRQ_set_mask(uint8_t IRQline);
 void IRQ_clear_mask(uint8_t IRQline);
